refactor: 스타일 파일 네이밍에서 style로 된것 styles로 통일
