+++ frontend/src/components/common/Comment/Comment.tsx
@@ -1,21 +1,11 @@
 import { deleteComments } from '@/api/comments';
-import * as S from '@/components/common/Comment/Comment.style';
-import styled from '@emotion/styled';
+import * as S from '@/components/common/Comment/Comment.styles';
 import { useEffect, useState } from 'react';
 import { useMutation } from 'react-query';
 import CommentInputModal from '../CommentInputModal/CommentInputModal';
 import { queryClient } from '@/index';
 import { CommentType } from '@/types/commentResponse';
 import Loading from '@/components/common/Loading/Loading';
-export const DimmerContainer = styled.div`
-	position: fixed;
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	background-color: ${({ theme }) => theme.colors.GRAY_500};
-	z-index: 110;
-`;
 
 interface CommentProps extends CommentType {
 	articleId: string;
@@ -68,7 +58,7 @@ const Comment = ({ id, author, content, createdAt, isAuthor, articleId }: Commen
 
 			{isEditCommentOpen && (
 				<>
-					<DimmerContainer onClick={() => setIsEditCommentOpen(false)} />
+					<S.DimmerContainer onClick={() => setIsEditCommentOpen(false)} />
 					<CommentInputModal
 						closeModal={() => setIsEditCommentOpen(false)}
 						articleId={String(articleId)}
