+++ frontend/src/components/common/SearchBar/SearchBar.styles.tsx
