+++ frontend/src/components/common/SearchBar/SearchBar.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import * as S from '@/components/common/SearchBar/SearchBar.style';
+import * as S from '@/components/common/SearchBar/SearchBar.styles';
 
 const SearchBar = () => {
 	const [search, setSearch] = useState('');
