+++ frontend/src/components/layout/TabBar/TabBar.tsx
@@ -1,5 +1,5 @@
 import MenuSlider from '@/components/common/MenuSlider/MenuSlider';
-import * as S from '@/components/layout/TabBar/TabBar.style';
+import * as S from '@/components/layout/TabBar/TabBar.styles';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
