+++ frontend/src/pages/CategorySelector/CategorySelector.styles.tsx
