+++ frontend/src/pages/Detail/index.styles.tsx
