+++ frontend/src/pages/Detail/index.tsx
@@ -2,7 +2,7 @@ import ArticleContent from '@/components/common/ArticleContent/ArticleContent';
 import Comment from '@/components/common/Comment/Comment';
 import CommentInputModal from '@/components/common/CommentInputModal/CommentInputModal';
 
-import * as S from '@/pages/Detail/index.style';
+import * as S from '@/pages/Detail/index.styles';
 import { ArticleType } from '@/types/articleResponse';
 import { CommentType } from '@/types/commentResponse';
 import React, { useState } from 'react';
