+++ frontend/src/pages/Login/LoginButton/LoginButton.tsx
@@ -1,5 +1,5 @@
 import { LOGIN_TYPE } from '@/constants/loginType';
-import * as S from '@/pages/Login/LoginButton/LoginButton.style';
+import * as S from '@/pages/Login/LoginButton/LoginButton.styles';
 
 export interface LoginButtonProps {
 	loginType: keyof typeof LOGIN_TYPE;
